Fixed Jira module into one and added a dropdown list for new_issue
