+++ .env
@@ -61,3 +61,4 @@ IRC_TARGET=
 # Jira
 JIRA_URL=
 JIRA_TOKEN=
+JIRA_ISSUE_TYPE_LIST=
