+++ .test.env
@@ -42,5 +42,6 @@ IRC_PASSWORD=
 IRC_TARGET=#notify-test
 
 # Jira
+JIRA_ISSUE_TYPE_LIST=["Bug", "Task", "Story", "Epic"]
 JIRA_URL=https://jira.atlassian.com
 JIRA_TOKEN=123456
