+++ app/tests/api/api_v1/test_jira_comment.py
@@ -5,22 +5,21 @@ from fastapi.testclient import TestClient
 
 from app.core.config import settings
 
-options = {'server': settings.JIRA_URL}
 
 @mock.patch('jira.JIRA.add_comment')
 def test_jira_comment_default(
         mock_post, client: TestClient, api_key_headers: Dict[str, str]) -> None:
-    _comment_key = 'JRA-9'
+    _issue_key = 'JRA-9'
     params = {
-        "comment_key": _comment_key
+        "issue_key": _issue_key
     }
     _body = {
         "body": "SAMPLE MESSAGE"
     }
-    api_key_headers.update({"url": settings.JIRA_URL})
+    api_key_headers.update({"url": "https://jira.atlassian.com"})
 
 
-    r = client.post(f"{settings.API_V1_STR}/jira_add_comment",
+    r = client.post(f"{settings.API_V1_STR}/jira/add_comment",
                     json=_body,
                     params=params,
                     allow_redirects=True,
